[REF] web,*: rework CalendarFilterPanel
*hr_holidays

This commit reworks the CalendarFilterPanel component. This
component was badly named: it wasn't a panel, it was simply a
t-foreach on filter sections, that was used inside the calendar
side panel.

We thus remove it, and introduce instead a CalendarFilterSection
component, which is dedicated to a single section (the t-foreach
is thus moved up to the parent).

This component structure is more intuitive, easy to understand.

Followup of task-4510549
