+++ addons/hr_holidays/static/src/views/calendar/filter_section/calendar_filter_section.dark.scss
