+++ addons/hr_holidays/static/src/views/calendar/filter_section/calendar_filter_section.js
@@ -1,16 +1,16 @@
-import { CalendarFilterPanel } from "@web/views/calendar/filter_panel/calendar_filter_panel";
+import { CalendarFilterSection } from "@web/views/calendar/calendar_filter_section/calendar_filter_section";
 import { TimeOffCardMobile } from "../../../dashboard/time_off_card";
 
 import { useService } from "@web/core/utils/hooks";
 import { useState, onWillStart } from "@odoo/owl";
 
-export class TimeOffCalendarFilterPanel extends CalendarFilterPanel {
+export class TimeOffCalendarFilterSection extends CalendarFilterSection {
     static components = {
-        ...CalendarFilterPanel.components,
+        ...CalendarFilterSection.components,
         TimeOffCardMobile,
     };
     static subTemplates = {
-        filter: "hr_holidays.CalendarFilterPanel.filter",
+        filter: "hr_holidays.CalendarFilterSection.filter",
     };
 
     setup() {
@@ -20,18 +20,15 @@ export class TimeOffCalendarFilterPanel extends CalendarFilterPanel {
         this.leaveState = useState({
             holidays: [],
         });
-
         onWillStart(this.loadFilterData);
     }
 
     async loadFilterData() {
-        if (!this.env.isSmall) {
+        if (!this.env.isSmall || this.section.fieldName !== "holiday_status_id") {
             return;
         }
-
         const filterData = {};
         const data = await this.orm.call("hr.leave.type", "get_allocation_data_request", []);
-
         data.forEach((leave) => {
             filterData[leave[3]] = leave;
         });
