+++ addons/project/static/src/views/project_task_calendar/project_task_calendar_controller.js
@@ -1,12 +1,12 @@
 import { _t } from "@web/core/l10n/translation";
 import { CalendarController } from "@web/views/calendar/calendar_controller";
 import { subTaskDeleteConfirmationMessage } from "@project/views/project_task_form/project_task_form_controller";
-import { ProjectTaskCalendarFilterPanel } from "./project_task_calendar_filter_panel/project_task_calendar_filter_panel";
+import { ProjectTaskCalendarSidePanel } from "./project_task_calendar_side_panel";
 
 export class ProjectTaskCalendarController extends CalendarController {
     static components = {
         ...ProjectTaskCalendarController.components,
-        FilterPanel: ProjectTaskCalendarFilterPanel,
+        CalendarSidePanel: ProjectTaskCalendarSidePanel,
     };
 
     get editRecordDefaultDisplayText() {
