--- addons/project/static/src/views/project_task_calendar/project_task_calendar_filter_panel/project_task_calendar_filter_panel.js
@@ -1,7 +1,0 @@
-import { CalendarFilterPanel } from "@web/views/calendar/filter_panel/calendar_filter_panel";
-
-export class ProjectTaskCalendarFilterPanel extends CalendarFilterPanel { 
-    static subTemplates = {
-        filter: "project.ProjectTaskCalendarFilterPanel.filter",
-    };
-}
