+++ addons/project/static/src/views/project_task_calendar/project_task_calendar_filter_section/project_task_calendar_filter_section.js
@@ -1,0 +1,7 @@
+import { CalendarFilterSection } from "@web/views/calendar/calendar_filter_section/calendar_filter_section";
+
+export class ProjectTaskCalendarFilterSection extends CalendarFilterSection { 
+    static subTemplates = {
+        filter: "project.ProjectTaskCalendarFilterSection.filter",
+    };
+}
