+++ addons/project/static/src/views/project_task_calendar/project_task_calendar_filter_section/project_task_calendar_filter_section.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-    <t t-name="project.ProjectTaskCalendarFilterPanel.filter" t-inherit="web.CalendarFilterPanel.filter" t-inherit-mode="primary">
+    <t t-name="project.ProjectTaskCalendarFilterSection.filter" t-inherit="web.CalendarFilterSection.filter" t-inherit-mode="primary">
         <xpath expr="//span[@t-esc='filter.label']" position="before">
             <span t-if="props.model.isCheckProject and !filter.value" class="text-danger pe-1">
                 <i class="fa fa-lock"/>
