+++ addons/web/static/src/views/calendar/calendar_filter_section/calendar_filter_section.scss
