+++ addons/web/static/src/views/calendar/mobile_filter_panel/calendar_mobile_filter_panel.xml
@@ -6,7 +6,6 @@
             <div class="o_filter me-auto d-flex overflow-auto">
                 <t t-foreach="props.model.filterSections" t-as="section" t-key="section.fieldName">
                     <t t-if="section.filters.length gt 0">
-
                         <t t-foreach="getSortedFilters(section)" t-as="filter" t-key="filter.value">
                             <span t-if="filter.active" class="d-flex align-items-center text-nowrap ms-1 me-2">
                                 <i class="fa fa-fw" t-att-class="getFilterColor(filter) !== 'o_color_false' ? 'fa-circle ' + getFilterColor(filter) : 'fa-circle-o'"/>
